Move listApplicableTablePrivileges to SqlStandardAccessControl
