+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/metastore/thrift/ThriftMetastoreUtil.java
@@ -22,7 +22,6 @@ import com.google.common.primitives.Shorts;
 import io.trino.plugin.hive.HiveBasicStatistics;
 import io.trino.plugin.hive.HiveBucketProperty;
 import io.trino.plugin.hive.HiveType;
-import io.trino.plugin.hive.authentication.HiveIdentity;
 import io.trino.plugin.hive.metastore.Column;
 import io.trino.plugin.hive.metastore.Database;
 import io.trino.plugin.hive.metastore.HiveColumnStatistics;
@@ -284,22 +283,6 @@ public final class ThriftMetastoreUtil
                 Stream.of(new HivePrincipal(USER, identity.getUser())),
                 listEnabledRoles(identity, metastore::listRoleGrants)
                         .map(role -> new HivePrincipal(ROLE, role)));
-    }
-
-    public static Stream<HivePrivilegeInfo> listApplicableTablePrivileges(SemiTransactionalHiveMetastore metastore, String databaseName, String tableName, ConnectorIdentity identity)
-    {
-        String user = identity.getUser();
-        HivePrincipal userPrincipal = new HivePrincipal(USER, user);
-        Stream<HivePrincipal> principals = Stream.concat(
-                Stream.of(userPrincipal),
-                listApplicableRoles(metastore, userPrincipal)
-                        .map(role -> new HivePrincipal(ROLE, role)));
-        return listTablePrivileges(metastore, new HiveIdentity(identity), databaseName, tableName, principals);
-    }
-
-    private static Stream<HivePrivilegeInfo> listTablePrivileges(SemiTransactionalHiveMetastore metastore, HiveIdentity identity, String databaseName, String tableName, Stream<HivePrincipal> principals)
-    {
-        return principals.flatMap(principal -> metastore.listTablePrivileges(identity, databaseName, tableName, Optional.of(principal)).stream());
     }
 
     public static boolean isRoleEnabled(ConnectorIdentity identity, Function<HivePrincipal, Set<RoleGrant>> listRoleGrants, String role)
